Move inbox rules into their own section in settings
Closes #11075
