+++ src/ui/base/icons/Icons.ts
@@ -104,6 +104,8 @@ export const enum Icons {
 	FolderOutline = "FolderOutline",
 	FormatClear = "FormatClear",
 	Free = "Free",
+	FunnelFilled = "FunnelFilled",
+	FunnelOutline = "FunnelOutline",
 	GearWheelFilled = "GearWheelFilled",
 	GearWheelOutline = "GearWheelOutline",
 	GenericLockFilled = "GenericLockFilled",
@@ -406,6 +408,8 @@ export const IconsSvg: Record<Icons, string> = Object.freeze({
 	FolderOutline: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M19.102 8a1.9 1.9 0 0 1 1.395.625 1.99 1.99 0 0 1 .497 1.472v.01l-.643 6.662a2.3 2.3 0 0 1-.659 1.579l-.162.147c-.336.278-.745.45-1.176.494l-.216.011H4.858l-.216-.011a2.2 2.2 0 0 1-1.337-.642 2.3 2.3 0 0 1-.66-1.578l-.639-6.662v-.01A2 2 0 0 1 2.05 9.5l.055-.193q.097-.286.273-.529l.126-.153A1.87 1.87 0 0 1 3.9 8zM3.767 9.291a.64.64 0 0 0-.345.2.67.67 0 0 0-.17.502l.003.038.634 6.614.005.052v.053c.005.26.108.507.289.689.18.181.424.284.677.284h13.277a.96.96 0 0 0 .677-.284 1 1 0 0 0 .288-.689v-.053l.006-.053.64-6.65-.003-.137a.67.67 0 0 0-.167-.366.65.65 0 0 0-.345-.2l-.133-.014H3.9z"/><mask id="a"><path d="M4 8.534V5.62c0-.43.168-.841.466-1.145A1.58 1.58 0 0 1 5.59 4h3.018c.314 0 .62.095.882.272l1.107.751c.261.178.568.272.882.272h6.198c.422 0 .827.171 1.125.475s.466.715.466 1.145v1.62"/></mask><path d="M2.75 8.534a1.25 1.25 0 1 0 2.5 0zM4 5.62H2.75zM8.608 4V2.75zm.882.272-.702 1.034zm1.107.751.702-1.034zm.882.272v1.25zm6.198 0v-1.25zm.34 3.24a1.25 1.25 0 0 0 2.5 0zM4 8.534h1.25V5.618h-2.5v2.915zm0-2.916h1.25c0-.105.041-.201.108-.269l-.892-.876-.892-.876A2.88 2.88 0 0 0 2.75 5.62H4Zm.466-1.145.892.876c.065-.067.15-.1.232-.1v-2.5c-.76 0-1.485.308-2.016.848zM5.59 4v1.25h3.018v-2.5H5.59zm3.018 0v1.25a.3.3 0 0 1 .18.056l.702-1.034.702-1.034a2.8 2.8 0 0 0-1.584-.488zm.882.272-.702 1.034 1.107.752.702-1.035.702-1.034-1.107-.751zm1.107.751-.702 1.035a2.8 2.8 0 0 0 1.584.487v-2.5a.3.3 0 0 1-.18-.056zm.882.272v1.25h6.198v-2.5H11.48v1.25Zm6.198 0v1.25c.083 0 .168.034.233.1l.892-.875.892-.876a2.83 2.83 0 0 0-2.017-.849zm1.125.475-.892.876a.38.38 0 0 1 .108.269h2.5c0-.754-.294-1.481-.824-2.021zm.466 1.145h-1.25v1.62h2.5v-1.62z" mask="url(#a)"/></svg>`,
 	FormatClear: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M18.158 17.952c.216.215.232.548.037.743s-.528.18-.744-.036L3.89 5.097c-.216-.215-.232-.548-.037-.743.195-.196.528-.18.744.036zm-7.445-4.279-2.53 5.818a.847.847 0 0 1-1.552-.675l2.795-6.428zM19.578 4a.923.923 0 0 1 0 1.847h-5.462l-1.904 4.378-1.287-1.286 1.345-3.092H7.827l-1.34-1.338c-.207-.205-.333-.332-.476-.41A.9.9 0 0 1 6.424 4z"/></svg>`,
 	Free: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12.002 5.24c1.428 0 2.59-1.175 2.59-2.62S13.43 0 12.002 0 9.413 1.175 9.413 2.62s1.162 2.62 2.59 2.62Zm0-3.743c.612 0 1.11.504 1.11 1.123 0 .62-.498 1.123-1.11 1.123-.611 0-1.11-.504-1.11-1.123s.499-1.123 1.11-1.123"/><path d="M21.66 4.2a1.79 1.79 0 0 0-1.063-1.555 1.79 1.79 0 0 0-1.893.265C16.518 4.742 13.61 6.7 12.002 6.7c-1.427 0-4.554-2.077-6.799-3.855a1.75 1.75 0 0 0-1.889-.194 1.76 1.76 0 0 0-.973 1.691c.035.824.59 1.808 1.322 2.346 2.286 1.74 4.817 3.76 5.244 4.247.156.517.013 1.434-.04 1.764l-.269 1.37-1.734 7.69c-.108.459-.016.96.254 1.377.286.442.74.746 1.243.835.935.166 1.804-.42 2.115-1.425l1.527-4.775 1.53 4.778c.27.89.985 1.45 1.8 1.45a1.88 1.88 0 0 0 1.548-.83 1.82 1.82 0 0 0 .268-1.395l-.002-.011-1.736-7.697-.27-1.37c-.176-1.121-.064-1.562-.002-1.705.386-.424 2.932-2.491 5.24-4.3.852-.584 1.284-1.388 1.284-2.389zm-2.123 1.26-.043.031c-.013.01-1.32 1.035-2.639 2.098-2.76 2.228-2.896 2.462-2.992 2.64-.323.59-.38 1.454-.18 2.718l.005.028.275 1.393.004.02 1.736 7.701a.32.32 0 0 1-.048.243.4.4 0 0 1-.258.165c-.307.055-.438-.346-.451-.392l-.004-.012-1.745-5.452a11 11 0 0 1-.208-.74l-.268-1.105a.74.74 0 0 0-.719-.57.74.74 0 0 0-.718.57l-.247 1.02a13 13 0 0 1-.254.91l-1.715 5.363-.003.008c-.032.106-.161.45-.449.4a.4.4 0 0 1-.26-.182.3.3 0 0 1-.053-.21l.003-.009 1.737-7.704.004-.02.275-1.394.005-.027c.128-.812.237-1.843-.064-2.626-.093-.245-.19-.498-3.039-2.757a194 194 0 0 0-2.677-2.08q-.006-.004-.011-.009C3.309 4.581 3.9 4.028 3.97 3.993a.29.29 0 0 1 .321.03 40 40 0 0 0 3.366 2.4c1.9 1.194 3.32 1.775 4.345 1.775 1.11 0 2.588-.583 4.394-1.731a32 32 0 0 0 3.251-2.404.33.33 0 0 1 .354-.047c.065.03.705.443-.464 1.445Z"/></svg>`,
+	FunnelFilled: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M13.5 20a.9.9 0 0 1-.284-.047l-3.006-1.026a.9.9 0 0 1-.443-.337.94.94 0 0 1-.168-.54v-4.584l-.003-.005L3.27 5.905a1.18 1.18 0 0 1-.168-1.234A1.12 1.12 0 0 1 4.127 4h15.748c.215 0 .426.063.608.182.18.12.325.289.415.49a1.18 1.18 0 0 1-.168 1.233l-6.327 7.556-.003.005v5.61a.94.94 0 0 1-.263.654.9.9 0 0 1-.637.27"/></svg>`,
+	FunnelOutline: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M19.891 4a1.1 1.1 0 0 1 .467.105l.14.078.126.098q.12.107.204.24l.074.14.054.148a1.14 1.14 0 0 1-.125.906l-.093.128-6.334 7.592-.004.005a.06.06 0 0 0-.012.035v5.633a.89.89 0 0 1-.356.717.88.88 0 0 1-.813.127l-3.01-1.03h-.003a.88.88 0 0 1-.594-.847v-4.602l-.012-.035-.002-.003L3.26 5.843A1.13 1.13 0 0 1 3.1 4.66l.075-.14a1.1 1.1 0 0 1 .33-.338l.14-.078c.146-.067.305-.102.467-.102zm-9.395 8.693.065.083c.142.202.218.443.22.689v4.427l2.438.833v-5.262a1.2 1.2 0 0 1 .289-.773l6.284-7.53H4.208z"/></svg>`,
 	GearWheelFilled: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M18.462 12c0-.996.617-1.783 1.538-2.325a8.5 8.5 0 0 0-.658-1.592c-1.038.271-1.875-.133-2.584-.841-.704-.705-.92-1.546-.65-2.584A8.4 8.4 0 0 0 14.517 4C13.975 4.925 13 5.537 12 5.537c-.996 0-1.975-.612-2.517-1.537a8.4 8.4 0 0 0-1.591.658c.27 1.038.054 1.875-.65 2.584-.705.704-1.546 1.112-2.58.841A8 8 0 0 0 4 9.675c.925.542 1.542 1.33 1.542 2.325S4.925 13.975 4 14.517c.167.558.388 1.091.658 1.591 1.038-.27 1.875-.054 2.58.65.708.704.92 1.546.65 2.584q.755.406 1.591.658c.542-.925 1.521-1.538 2.517-1.538s1.975.613 2.516 1.538a8.4 8.4 0 0 0 1.592-.658c-.27-1.038-.054-1.875.65-2.584.704-.704 1.546-1.112 2.583-.841a8.6 8.6 0 0 0 .659-1.592c-.921-.542-1.534-1.33-1.534-2.325M12 16c-2.204 0-4-1.792-4-4a4.001 4.001 0 0 1 8 0c0 2.208-1.792 4-4 4"/></svg>`,
 	GearWheelOutline: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M17.433 12c0-1.045.555-1.86 1.32-2.434a7 7 0 0 0-.225-.545c-1.023.09-1.857-.386-2.51-1.04a3 3 0 0 1-.811-1.382 3 3 0 0 1-.059-1.133 7 7 0 0 0-.544-.225c-.634.798-1.588 1.325-2.605 1.325-1.013 0-1.971-.526-2.605-1.325a7 7 0 0 0-.545.225c.049.388.035.766-.058 1.132-.133.525-.413.983-.81 1.381v.002c-.65.65-1.487 1.13-2.51 1.04q-.125.266-.228.545c.77.575 1.326 1.389 1.326 2.434 0 1.014-.53 1.97-1.329 2.604q.102.28.225.545a3 3 0 0 1 1.13.059c.526.133.984.413 1.382.81h-.001c.4.399.68.858.813 1.384.092.366.106.744.057 1.131q.267.123.544.225c.634-.798 1.592-1.324 2.605-1.324s1.97.526 2.605 1.324q.278-.102.545-.225a3 3 0 0 1 .058-1.13c.133-.526.413-.983.81-1.383v-.001c.65-.65 1.487-1.129 2.51-1.04q.124-.27.226-.547c-.765-.575-1.316-1.39-1.316-2.432m-2.309 0a3.126 3.126 0 0 0-6.25 0 3.126 3.126 0 0 0 6.25 0m1.25 0a4.376 4.376 0 0 1-8.75 0 4.376 4.376 0 0 1 8.75 0m2.309 0c0 .637.382 1.201 1.128 1.64l.424.25-.141.47a8.7 8.7 0 0 1-.665 1.608l-.233.432-.476-.124c-.685-.178-1.256.062-1.822.627-.263.264-.414.53-.484.807-.07.278-.069.609.037 1.015l.124.474-.43.234a8.5 8.5 0 0 1-1.612.666l-.47.14-.248-.424c-.41-.701-1.137-1.131-1.82-1.131s-1.41.43-1.82 1.131l-.248.424-.47-.14a8.6 8.6 0 0 1-1.61-.665l-.432-.233.124-.476c.106-.407.108-.74.038-1.018-.07-.276-.22-.541-.483-.803h-.002c-.263-.263-.527-.415-.803-.485-.278-.07-.608-.069-1.015.037l-.474.124-.234-.43a8.5 8.5 0 0 1-.666-1.612l-.14-.47.424-.248c.703-.412 1.135-1.139 1.135-1.82 0-.636-.383-1.2-1.135-1.64l-.424-.249.14-.47a8.2 8.2 0 0 1 .672-1.614l.236-.427.47.124c.68.178 1.252-.062 1.818-.627.262-.265.414-.53.484-.807.07-.279.069-.609-.037-1.015l-.124-.474.43-.234a8.5 8.5 0 0 1 1.612-.666l.47-.14.248.424c.41.701 1.137 1.131 1.82 1.131.686 0 1.409-.43 1.82-1.131l.248-.424.47.14a8.5 8.5 0 0 1 1.611.666l.431.234-.124.474c-.106.406-.108.738-.037 1.017.07.277.222.543.484.805.569.568 1.136.805 1.822.627l.474-.124.234.43c.275.508.498 1.052.666 1.61l.141.47-.424.25c-.748.44-1.132 1.004-1.132 1.64"/></svg>`,
 	GenericLockFilled: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M16.136 9.867h-.59V7.2c0-.849-.374-1.663-1.039-2.263S12.94 4 12 4s-1.842.337-2.507.937S8.455 6.351 8.455 7.2v2.667h-.591c-.627 0-1.228.225-1.67.625-.444.4-.693.942-.694 1.508v5.867c0 .565.25 1.107.693 1.508.443.4 1.044.624 1.67.625h8.273c.627 0 1.228-.226 1.671-.625.443-.4.692-.943.693-1.508V12c0-.566-.25-1.108-.693-1.508a2.5 2.5 0 0 0-1.67-.625Zm-1.978 0H9.842V7.2c0-.566.25-1.028.692-1.428.444-.4.84-.505 1.466-.505.627 0 1.022.104 1.466.505.443.4.692.862.692 1.428z"/></svg>`,
